--- allOrders/allOrders.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="/styles/styles.css">
-    <title>Fast Food Receipt Generator</title>
-</head>
-<body>
-    <div class="container">
-        <h1>Accepterede ordre</h1>
-
-        <!-- Form elements -->
-            <table>
-                <tbody id="receipt-data">
-                    <!-- Receipt data will be populated by JavaScript -->
-                </tbody>
-            </table>
-        </div>
-    </div>
-    <script src="allOrders.js"></script>
-</body>
-</html>
