+++ index.html
@@ -11,6 +11,7 @@
 
     <link rel="stylesheet" href="styles/meny-styles-non-responsive.css" />
     <link rel="stylesheet" href="styles/pizza-management.css" />
+    <link rel="stylesheet" href="styles/styles.css">
     <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500&display=swap" rel="stylesheet">
